VACMS-20493: Moves staffProfile to products folder
