+++ src/data/queries/index.ts
@@ -27,7 +27,7 @@ import * as MediaVideo from '@/data/queries/mediaVideo'
 import * as NewsStory from '@/data/queries/newsStory'
 import * as NewsStoryTeaser from '@/data/queries/newsStoryTeaser'
 import * as NumberCallout from '@/data/queries/numberCallout'
-import * as PersonProfile from '@/data/queries/personProfile'
+import * as PersonProfile from '@/products/staffProfile/query'
 import * as PhoneNumber from '@/data/queries/phoneNumber'
 import * as PressRelease from '@/data/queries/pressRelease'
 import * as PressReleaseListing from '@/data/queries/pressReleaseListing'
